Implement splash screen loading and display in fusee stage2
Splash loader will look for a file called `splash.bmp` on the root of the SD card and display it. If the file does not exist, or it can't be read for whatever reason,
it will display the default splash linked into the binary.

Stage2 binary is now quite large (>3MB) as a result of statically linking in the default splash screen.