docs: Add a bunch of 0.11.2's changelog ahead of time.
0.11.2 will release sometime in the next week or two or so.

I am writing this now while I remember all the things that have happened,
so that I don't forget about them when release comes.

There is also a PR for dmnt changes that will be merged into 0.11.2 not included here,
and there may be more changes before the release occurs.