fusee-primary: tweak sdram_lp0_save_params to save code space
This code is essentially a giant unrolled loop over an array of register
copies. I've manually re-rolled the loop so that rather than emitting code for
each parameter to read it, perform bit arithmetic on it, and write it, there is
instead a static list of parameters which can be encoded using much less memory.